+++ lib/axn/webhooks/inbound.rb
@@ -2,6 +2,7 @@
 
 require_relative "inbound/dsl"
 require_relative "inbound/endpoint"
+require_relative "inbound/router"
 
 module Axn
   module Webhooks
@@ -25,7 +26,7 @@ module Axn
 
       dsl = Inbound::DSL.new
       dsl.instance_exec(&block)
-      Inbound.register(name, Inbound::Endpoint.new(name:, verifier: dsl.__verifier__))
+      Inbound.register(name, Inbound::Endpoint.new(name:, verifier: dsl.__verifier__, dispatch: dsl.__dispatch__))
     end
   end
 end